Move image caching helpers to the lib
